Refactored groups code
